Explicit instruction (reminder) branch change
Added to the instructions to change over to the develop branch as this always slips my mind.  In theory, people working in the develop branch should know how to problem solve this, but it's a good reminder.  We will need to edit this out or change it as develop merges into master.